Add modularization
