+++ src/about.js
@@ -1,0 +1,8 @@
+// ABOUT
+const aboutMe = () => {
+  const about = document.createElement('div');
+  about.classList.add('about');
+  return about;
+};
+
+export default aboutMe;
